Move RHVoice_message_type, RHVoice_punctuation_mode, RHVoice_capitals_mode
